refactor: tweaks for EmbedCodeDialog
1. Update the styles applied to the code block.
2. Move the component to 'tsrc/components'.
3. Add a close button in the dialog.
